svn merge -r11143:HEAD svn://svn.valgrind.org/valgrind/branches/MACOSX106

There were some easy-to-resolve conflicts.

Then I had to fix up coregrind/link_tool_exe*.in -- those files had been
added independently on both the trunk and the branch, AFAICT.  I just
overwrote the trunk versions with the branch versions.



git-svn-id: svn://svn.valgrind.org/valgrind/trunk@11194

+++ drd/tests/tsan_unittest.cpp
@@ -6620,8 +6620,8 @@ void Getter() {
     }
     usleep(1000);
   }
-  printf("T=%d: non_zero_received=%d\n", 
-         (int)pthread_self(), non_zero_received);
+  printf("T=%ld: non_zero_received=%d\n", 
+         (size_t)pthread_self(), non_zero_received);
 } 
 
 void Run() {
